# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Ratinam Crackers Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserListResponse"
    post:
      operationId: createUser
      tags: [users]
      summary: Create user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /v1/users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ============================================================
  # LOCATIONS
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    # AUTH SCHEMAS
...
    User:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [SUPER_ADMIN, ERP_MANAGER, ACCOUNTANT, AGENT, WH_MANAGER, CASHIER]
        locationIds:
          type: array
          items:
            type: string
        email:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
    UserListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/User"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
    CreateUserBody:
      type: object
      required: [name, username, password, role]
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
        pin:
          type: string
        role:
          type: string
        email:
          type: string
        phone:
          type: string
        locationIds:
          type: array
          items:
            type: string
        maxDiscountPct:
          type: number
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
        locationIds:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        maxDiscountPct:
          type: number

    # PRODUCT SCHEMAS
...